refactor(client): restructure clients to decompose core client
