--- docs/api/modules/interface_ILocalizedString.md
@@ -1,25 +1,0 @@
-# Module: interface/OptionalPerLanguageMapping
-
-## Table of contents
-
-### Type aliases
-
-- [OptionalPerLanguageMapping](interface_ILocalizedString.md#ilocalizedstring)
-
-## Type aliases
-
-### OptionalPerLanguageMapping
-
-Ƭ **OptionalPerLanguageMapping**: { [key in Language]?: string}
-
-Supported languages for localized text.
-
-Corresponding to ISO 639-1:2002.
-
-These do not correspond to:
-Country code (ISO-3166) - e.g. Japan would be JP not ja.
-Square Enix labelling - eu=en, na=enUs, de=de, fr=fr, jp=ja
-
-#### Defined in
-
-[interface/OptionalPerLanguageMapping.ts:37](https://github.com/XIVStats/lodestone/blob/e81c3ec/src/interface/ILocalizedString.ts#L37)
